Rename tool-utils staging module to overlay
"staging.js" read as a deploy-tier name even though the module writes
production dist/ output and import maps. Rename to overlay.js to match
the exported overlayBuiltOutputs function and drop the ambiguity.
