+++ tools/utils/src/index.d.ts
@@ -1,5 +1,5 @@
 export * from './log.js';
+export * from './overlay.js';
 export * from './path.js';
 export * from './productionize.js';
 export * from './shell.js';
-export * from './staging.js';
