+++ tools/utils/src/overlay.d.ts
